Moved database into separate app for generalizing
