+++ app/groupformer/dbtools/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DbtoolsConfig(AppConfig):
+    name = 'dbtools'
