+++ app/groupformer/dbtools/tests.py
@@ -1,0 +1,20 @@
+from django.test import TestCase
+
+# Create your tests here.
+
+class DatabaseTests(TestCase):
+    def test_setup(self):
+        #Setup a basic structure
+        pass
+    
+    def test_adding(self):
+        #Add various additional entities and relationships
+        pass
+    
+    def test_removing(self):
+        #Remove different elements and test their outputs
+        pass
+    
+    def test_cleanup(self):
+        #Clear database by removing GroupFormers
+        pass
