+++ app/groupformer/project_input_test/tests.py
@@ -25,20 +25,3 @@ class ProjectIndexViewTests(TestCase):
         response = self.client.get(reverse('project_input_test:index'))
         self.assertContains(response, project.project_name)
         self.assertContains(response, project.project_description)
-
-class DatabaseTests(TestCase):
-    def test_setup(self):
-        #Setup a basic structure
-        pass
-    
-    def test_adding(self):
-        #Add various additional entities and relationships
-        pass
-    
-    def test_removing(self):
-        #Remove different elements and test their outputs
-        pass
-    
-    def test_cleanup(self):
-        #Clear database by removing GroupFormers
-        pass
